[chore] 코드 리뷰 - TripmateScaffold 적용 및 Map모듈 Mate모듈분리
