--- feature/map/src/main/java/com/tripmate/android/feature/map/viewmodel/MapUiEvent.kt
@@ -1,6 +1,0 @@
-package com.tripmate.android.feature.map.viewmodel
-
-
-sealed interface MapUiEvent {
-    data object ClickCurrentLocation : MapUiEvent
-}
