+++ feature/mate/build.gradle.kts
@@ -1,5 +1,8 @@
 @file:Suppress("INLINE_FROM_HIGHER_PLATFORM")
 
+import com.tripmate.android.implementation
+
+
 plugins {
     alias(libs.plugins.tripmate.android.feature)
 }
@@ -9,7 +12,12 @@ android {
 }
 
 dependencies {
+    implementation(project(":feature:map"))
     implementations(
         libs.timber,
+        libs.androidx.compose.material3,
+        libs.android.play.services.location,
+        libs.accompanist.permissions,
+        libs.kakao.maps
     )
 }
