+++ feature/mate/src/main/kotlin/com/tripmate/android/mate/viewmodel/MateUiEvent.kt
@@ -1,0 +1,6 @@
+package com.tripmate.android.mate.viewmodel
+
+
+sealed interface MateUiEvent {
+    data object ClickCurrentLocation : MateUiEvent
+}
